+++ driver/src/global/xaiegbl_regdef.h
@@ -584,7 +584,7 @@ typedef struct {
  */
 struct XAie_DmaMod {
 	u8  NumBds;
-	u8  NumLocks;
+	u16  NumLocks;
 	u8  ChIdxOffset;
 	u8  NumAddrDim;
 	u8  DoubleBuffering;
@@ -614,7 +614,7 @@ struct XAie_DmaMod {
 	AieRC (*SetIntrleave) (XAie_DmaDesc *Desc, u8 DoubleBuff,
 			u8 IntrleaveCount, u16 IntrleaveCurr);
 	AieRC (*SetMultiDim) (XAie_DmaDesc *Desc, XAie_DmaTensor *Tensor);
-	AieRC (*SetBdIter) (XAie_DmaDesc *Desc, u32 StepSize, u8 Wrap,
+	AieRC (*SetBdIter) (XAie_DmaDesc *Desc, u32 StepSize, u16 Wrap,
 			u8 IterCurr);
 	AieRC (*WriteBd)(XAie_DevInst *DevInst, XAie_DmaDesc *Desc,
 			XAie_LocType Loc, u8 BdNum);
@@ -628,9 +628,9 @@ struct XAie_DmaMod {
 			XAie_DmaDirection Dir, u32 TimeOutUs);
 	AieRC (*BdChValidity)(u8 BdNum, u8 ChNum);
 	AieRC (*UpdateBdLen)(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-			XAie_LocType Loc, u32 Len, u8 BdNum);
+			XAie_LocType Loc, u32 Len, u16 BdNum);
 	AieRC (*UpdateBdAddr)(XAie_DevInst *DevInst, const XAie_DmaMod *DmaMod,
-			XAie_LocType Loc, u64 Addr, u8 BdNum);
+			XAie_LocType Loc, u64 Addr, u16 BdNum);
 	AieRC (*GetChannelStatus)(XAie_DevInst *DevInst, XAie_LocType Loc,
 			const XAie_DmaMod *DmaMod, u8 ChNum,
 			XAie_DmaDirection Dir, u32 *Status);
@@ -744,7 +744,7 @@ typedef struct {
  * document.
  */
 struct XAie_LockMod {
-	u8  NumLocks;		/* Number of lock in the module */
+	u16  NumLocks;		/* Number of lock in the module */
 	s8  LockValUpperBound; 	/* Upper bound of the lock value */
 	s8  LockValLowerBound; 	/* Lower bound of the lock value */
 	u32 BaseAddr;		/* Base address of the lock module */
